Removed 'using System;' and added Dust.Math

+++ Dust/Emission/BoxEmitter.cs
@@ -14,7 +14,6 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
-using System;
 
 namespace GameClay.Dust
 {
@@ -361,7 +360,7 @@ namespace GameClay.Dust
                 // TODO: Matrix and transform stuff
 
                 // Length
-                float len = (float)Math.Sqrt((posX * posX) + (posY * posY) + (posZ * posZ));
+                float len = Math.Sqrt((posX * posX) + (posY * posY) + (posZ * posZ));
 
                 // Normalize
                 float velX = posX / len;
